#39 allow to check init & unlock
